fix: extract storybook
