+++ apps/main/package.json
@@ -13,9 +13,7 @@
     "g:gql": "graphql-codegen --config ./scripts/graphql-codegen/server-codegen.yml && graphql-codegen --config ./scripts/graphql-codegen/client-codegen.yml",
     "g:gql:watch": "concurrently 'graphql-codegen --watch --config ./scripts/graphql-codegen/server-codegen.yml' 'graphql-codegen --config ./scripts/graphql-codegen/client-codegen.yml --watch'",
     "ut": "DEBUG_PRINT_LIMIT=10000 jest --silent=false",
-    "ut:coverage": "DEBUG_PRINT_LIMIT=10000 jest --silent=false --coverage",
-    "sb": "start-storybook -p 6006 -s ./public",
-    "build-sb": "build-storybook -s ./public"
+    "ut:coverage": "DEBUG_PRINT_LIMIT=10000 jest --silent=false --coverage"
   },
   "dependencies": {
     "@geist-ui/react-icons": "1.0.1",
@@ -25,6 +23,7 @@
     "@react-aria/button": "3.4.3",
     "@sendinblue/client": "3.1.0",
     "@tailwindcss/typography": "0.4.1",
+    "@tiptap/core": "2.0.0-beta.175",
     "@tiptap/extension-link": "2.0.0-beta.37",
     "@tiptap/extension-placeholder": "2.0.0-beta.48",
     "@tiptap/extension-underline": "2.0.0-beta.23",
@@ -88,14 +87,6 @@
     "@graphql-codegen/typescript-urql": "3.5.7",
     "@graphql-typed-document-node/core": "3.1.1",
     "@relative-ci/agent": "3.1.2",
-    "@storybook/addon-essentials": "6.4.22",
-    "@storybook/addon-interactions": "6.4.22",
-    "@storybook/addon-links": "6.4.22",
-    "@storybook/builder-webpack5": "6.4.22",
-    "@storybook/manager-webpack5": "6.4.22",
-    "@storybook/react": "6.4.22",
-    "@storybook/testing-library": "0.0.11",
-    "@storybook/testing-react": "1.2.4",
     "@testing-library/jest-dom": "5.16.4",
     "@testing-library/react": "13.1.1",
     "@testing-library/user-event": "14.1.1",
@@ -115,12 +106,10 @@
     "@types/tailwindcss": "3.0.10",
     "@types/web-push": "3.3.2",
     "@types/ws": "8.5.3",
-    "@typescript-eslint/eslint-plugin": "5.21.0",
     "@urql/devtools": "2.0.3",
     "@urql/storybook-addon": "1.0.8",
+    "@storybook/testing-react": "1.2.4",
     "color-convert": "2.0.1",
-    "concurrently": "7.1.0",
-    "cypress": "9.2.0",
     "dotenv": "16.0.0",
     "dotenv-cli": "5.1.0",
     "dotenv-expand": "8.0.3",
@@ -132,11 +121,7 @@
     "postcss-flexbugs-fixes": "5.0.2",
     "postcss-preset-env": "7.4.4",
     "sass": "1.51.0",
-    "sass-loader": "12.6.0",
-    "storybook-addon-next": "1.6.2",
-    "tsconfig-paths-webpack-plugin": "3.5.2",
     "typescript": "4.6.4",
-    "webpack": "5.72.0",
     "whatwg-fetch": "3.6.2"
   }
 }
