+++ apps/main/src/__tests__/jest.setup.ts
@@ -4,7 +4,7 @@ import DotEnv from 'dotenv';
 import 'intersection-observer';
 import 'whatwg-fetch';
 
-import * as globalStorybookConfig from '../../.storybook/preview';
+import * as globalStorybookConfig from '../../../storybook/.storybook/preview';
 import { server } from './fixtures/server';
 import './mocks/mock-use-current-user';
 import { cleanEvents } from './mocks/next-router';
