+++ apps/storybook/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@chirpy/storybook",
+  "private": true,
+  "scripts": {
+    "sb": "start-storybook -p 6006 -s ../main/public",
+    "build-sb": "build-storybook -s ../main/public"
+  },
+  "devDependencies": {
+    "@babel/core": "7.16.10",
+    "@storybook/addon-actions": "6.4.22",
+    "@storybook/addon-essentials": "6.4.22",
+    "@storybook/addon-interactions": "6.4.22",
+    "@storybook/addon-links": "6.4.22",
+    "@storybook/builder-webpack5": "6.4.22",
+    "@storybook/manager-webpack5": "6.4.22",
+    "@storybook/react": "6.4.22",
+    "@storybook/testing-library": "0.0.11",
+    "babel-loader": "8.2.5",
+    "css-loader": "6.7.1",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "resolve-url-loader": "4.0.0",
+    "next": "12.1.5",
+    "postcss": "8.4.13",
+    "postcss-flexbugs-fixes": "5.0.2",
+    "postcss-preset-env": "7.4.4",
+    "postcss-loader": "6.2.1",
+    "sass": "1.51.0",
+    "sass-loader": "12.6.0",
+    "style-loader": "3.3.1",
+    "storybook-addon-next": "1.6.2",
+    "tailwindcss": "3.0.24",
+    "tsconfig-paths-webpack-plugin": "3.5.2",
+    "webpack": "5.72.0"
+  }
+}
