+++ apps/storybook/postcss.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../main/postcss.config');
