+++ package.json
@@ -24,6 +24,8 @@
     ]
   },
   "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "5.22.0",
+    "@typescript-eslint/parser": "5.22.0",
     "eslint": "8.14.0",
     "eslint-config-next": "12.1.5",
     "eslint-config-prettier": "8.5.0",
@@ -33,6 +35,7 @@
     "eslint-plugin-unicorn": "42.0.0",
     "react": "18.1.0",
     "simple-git-hooks": "2.7.0",
+    "next": "12.1.5",
     "lint-staged": "12.4.1",
     "prettier": "2.6.2",
     "turbo": "1.2.8",
